Move logger to top level package
