+++ audioswitch/src/main/java/com/twilio/audioswitch/Logger.kt
@@ -1,4 +1,4 @@
-package com.twilio.audioswitch.android
+package com.twilio.audioswitch
 
 import android.util.Log
