+++ audioswitch/src/main/java/com/twilio/audioswitch/selection/AudioDeviceSelector.kt
@@ -3,8 +3,8 @@ package com.twilio.audioswitch.selection
 import android.bluetooth.BluetoothAdapter
 import android.content.Context
 import android.media.AudioManager
+import com.twilio.audioswitch.Logger
 import com.twilio.audioswitch.android.BuildWrapper
-import com.twilio.audioswitch.android.Logger
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetConnectionListener
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetManager
 import com.twilio.audioswitch.selection.AudioDevice.BluetoothHeadset
