+++ audioswitch/src/test/java/com/twilio/audioswitch/wired/WiredHeadsetReceiverTest.kt
@@ -7,7 +7,7 @@ import com.nhaarman.mockitokotlin2.isA
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
-import com.twilio.audioswitch.android.Logger
+import com.twilio.audioswitch.Logger
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.nullValue
